requirements.txt
Django==1.8.6
gunicorn==19.0.0
# uWSGI==2.0.12
mysqlclient==1.3.7
dj-static==0.0.6
static3==0.5.1
boto==2.38.0
django-storages==1.1.8
djangorestframework==3.3.3
django-cors-headers==1.1.0
django-pylibmc==0.6.1
pylibmc==1.5.0
django-mailer==1.1
lockfile==0.12.2
django-rest-swagger==0.3.6